+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Upgrade to Flutter 3.7 & dependency updates
 - Fix album cover bug on currently playing page (#57)
 - Calculate album colors during library update and store them for better performance
+- Fix bug with smart lists in history entries
+- Added logging to files